requirements.txt
textbase-client>=0.1.4